fix decoder.
